Create get_tagged_item function

+++ likes/models.py
@@ -8,5 +8,5 @@ class LikedItems(models.Model):
     user = models.ForeignKey(User,on_delete=models.CASCADE)
 
     content_type = models.ForeignKey(ContentType,on_delete=models.CASCADE)
-    object_id = models.PositiveBigIntegerField()
+    object_id = models.PositiveIntegerField()
     content_object = GenericForeignKey() #retrive all the liked objects from a specific person
